### 2023-03-13 #71 @schackartk
[Re] Predictive model for the size of bubbles and droplets created in microfluidic T-junctions
**Original article:** van Steijn V., Kleijn C.R., Kreutzer M.T., Predictive model for the size of bubbles and droplets created in microfluidic T-junctions. Lab Chip, 2010,10, 2513-2518

Full text PDF available at: https://www.researchgate.net/publication/45114766_Predictive_model_for_the_size_of_bubbles_and_droplets_created_in_microfluidic_T-junctions

**Scientific domain:** Microfluidics, mathematical modeling
**Suggested editor:** Lorena Barba
**Suggested reviewers:** Vamsi Spandan and Kyle Niemeyer
