### 2023-03-27 6ab8fcb
test that fig_2b.png is created
