+++ article/content.tex
@@ -243,8 +243,8 @@ values as Figure O2a. However, six width ratios ($\frac{w_{in}}{w}$) are plotted
 This figure was again replicated using 1000 data points per width ratio. Channel width ($w$) was
 held constant while $h$ and $w_{in}$ were varied. As in the original work the corner roundness
 $\varepsilon=0$. The flow rate of the continuous phase was set as $q_c=1$, and the flow rate of the
-gutter was $q_{gutter}=0.1*q_c$ such that $\frac{q_{gutter}}{q_c}=0.1$, as described in the original
-work.
+gutter was calculated as $q_{gutter}=0.1*q_c$ such that $\frac{q_{gutter}}{q_c}=0.1$, as described
+in the original work.
 
 The squeezing coefficient was calculated based on these values using the function in the
 squeezing module, and plotted using the same axis scales as the original work.
@@ -347,8 +347,8 @@ $$
   \left(1-\frac{q_{gutter}}{q_c}\right)t\right)}}{2}
 \end{equation}
 
-\noindent Therefore $R$ can be evaluated as a function of time (taking the positive of the square root), and
-$2r$ can thus be calculated as:
+\noindent Therefore $R$ can be evaluated as a function of time (taking the positive of the square root) and
+$2r$ can then be calculated as:
 
 \begin{equation}
   2r = R - \sqrt{\left(R - w\right)^2 + \left(R - w_{in}\right)^2} + \varepsilon
@@ -357,7 +357,7 @@ $2r$ can thus be calculated as:
 \noindent Functions for both $R$ as a function of time, and $2r$ were added
 to the squeezing module.
 
-For replication of Figure O6, the provided geometries for $w$, $h$, and $q_c$ were used. 1000
+For replication of Figure O6, the provided geometries for $w$, $h$, and $q_c$ were used. 1000 data
 points for each inlet width ratio were generated spanning $\frac{q_c}{hw^2}t \in [0,10]$ by varying 
 $t$. Corner roundness ($\varepsilon$) was calculated based on $w$ as described. $2r$ was
 computed using the squeezing module and divided by $w$ for plotting. As in Figure O6, linetype
@@ -407,10 +407,10 @@ The figure generated using our modules (Figure \ref{fig2a}) is similar to
 Figure O2a in the original work. However, for $\frac{w_{in}}{w}>1$, the
 lines have a markedly more negative slope than those in Figure O2a.
 
-Interestingly however, a plot exactly matching Figure O2a could be achieved
+Interestingly however, a plot matching Figure O2a could be achieved
 (Figure \ref{fig2a_bad}). To do so,
 a separate function was written to calculate the dimensionless fill volume. In this function,
-the volume is not calculated using geometry, but instead used the equation in the original work,
+the volume is not calculated using geometry, but instead uses the equation in the original work,
 but neglecting a pair of parentheses.
 
 The portion of equation \eqref{nondim_fill_vol} that was modified is as follows.
@@ -419,7 +419,7 @@ The equation presented in the original work contains:
 $$
 \left(\left(\frac{\pi}{2} -
 \arcsin\left(1 - \frac{w}{w_{in}}\right)\right)\frac{w_{in}}{w} +
-\frac{\pi}{2}\right)\frac{h}{w}
+\frac{\pi}{2}\right)
 $$
 
 \noindent The rest of the equation was unmodified, however, the above portion was replaced with:
@@ -427,7 +427,7 @@ $$
 $$
 \left(\frac{\pi}{2} -
[…]
