+++ article/metadata.tex
@@ -11,7 +11,7 @@
 \def \reviewerIORCID{}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
-\def \dateRECEIVED{01 November 2018}
+\def \dateRECEIVED{13 March 2023}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
 \def \articleTITLE{[Re] Predictive model for the size of bubbles and droplets created in microfluidic T-junctions}
