+++ article/metadata.yaml
@@ -109,7 +109,7 @@ contributors:
 
 # This information will be provided by the editor
 dates:
-  - received:  November 1, 2018
+  - received:  March 13, 2023
   - accepted:
   - published: 
 
