### 2023-04-18 #71 @schackartk
Hi @labarba, thank you for your willingness to be the editor for our manuscript. We are not pressed for time and can be patient, but we don't wish to overburden you.
So, @rougier, if you have any other potential editors in mind we would be open to that as well. Whatever you see fit will work for us!
