### 2023-08-24 #71 @rougier
Sorry for delay and thanks for the reminder.

@ReScience/editors Dear editor, we would need someone to edit this submission that might be a bit out of your domain but we'll find reviewers.
@schackartk Could you suggest some reviewers (from the board of from outside)?
