### 2023-08-24 #71 @khinsen
@rougier This is close enough to my (former) field of work that I can handle this as an editor, but I wouldn't be able to propose suitable reviewers any more, as the people I might come up with are probably retired by now.
