### 2023-09-18 #71 @schackartk
I would suggest the following potential reviewers,

From the board:

- Vamsi Spandan
- Kyle Niemeyer
- Tarandeep Singh Kalra

Outside:
- Alexander Day (GitHub: aday913)
- Patarajarin Akarapipad (GitHub: patarajarina)
- Katie Sosnowski (GitHub: Katie-Sosnowski)

In the spirit of transparency, the suggested outside reviewers are former colleagues of the authors.
