### 2026-08-27 #71 @schackartk
Sure! I'd be interested in working on this.
